requirements.txt
requests
markdownify
pdfminer.six
PyYAML
pydantic>=2.0
pyarrow
